INSP: Convert some annotations into inspections
