+++ src/main/kotlin/org/rust/lang/core/types/Extensions.kt
@@ -5,9 +5,9 @@ import com.intellij.psi.util.CachedValueProvider
 import com.intellij.psi.util.CachedValuesManager
 import com.intellij.psi.util.PsiModificationTracker
 import org.rust.ide.utils.recursionGuard
-import org.rust.lang.core.psi.RsExpr
-import org.rust.lang.core.psi.RsTypeReference
+import org.rust.lang.core.psi.*
 import org.rust.lang.core.psi.ext.RsTypeBearingItemElement
+import org.rust.lang.core.psi.ext.ancestors
 import org.rust.lang.core.types.infer.inferDeclarationType
 import org.rust.lang.core.types.infer.inferExpressionType
 import org.rust.lang.core.types.infer.inferTypeReferenceType
@@ -18,6 +18,18 @@ import org.rust.lang.core.types.ty.TyUnknown
 val RsTypeReference.type: Ty
     get() = recursionGuard(this, Computable { inferTypeReferenceType(this) })
         ?: TyUnknown
+
+val RsTypeReference.lifetimeElidable: Boolean get() {
+    val typeOwner = topmostType.parent
+    return typeOwner !is RsFieldDecl && typeOwner !is RsTupleFieldDecl && typeOwner !is RsTypeAlias
+}
+
+val RsTypeReference.topmostType: RsTypeReference
+    get() = ancestors
+        .drop(1)
+        .filterNot { it is RsTypeArgumentList || it is RsPath }
+        .takeWhile { it is RsBaseType || it is RsTupleType || it is RsRefLikeType }
+        .lastOrNull() as? RsTypeReference ?: this
 
 val RsTypeBearingItemElement.type: Ty
     get() = CachedValuesManager.getCachedValue(this, CachedValueProvider {
