+++ src/main/resources/inspectionDescriptions/RsVariableMutable.html
@@ -1,5 +1,5 @@
 <html>
 <body>
-Inspect unused mut. This inspection can report false positives and is disabled by default.
+Detects unnecessary <code>mut</code> qualifiers. This inspection can report false positives and is disabled by default.
 </body>
 </html>
