+++ src/main/resources/inspectionDescriptions/RsWrongLifetimeParametersNumber.html
@@ -1,0 +1,8 @@
+<html>
+<body>
+Checks if the right number of lifetime parameters was used for a type or trait.
+
+Corresponds to <a href="https://doc.rust-lang.org/error-index.html#E0106">E0106</a> /
+<a href="https://doc.rust-lang.org/error-index.html#E0107">E0107</a> Rust errors.
+</body>
+</html>
