+++ src/main/resources/inspectionDescriptions/RsWrongTypeParametersNumber.html
@@ -1,0 +1,8 @@
+<html>
+<body>
+Checks if the right number of type parameters was used for a type or trait.
+
+Corresponds to <a href="https://doc.rust-lang.org/error-index.html#E0243">E0243</a> /
+<a href="https://doc.rust-lang.org/error-index.html#E0244">E0244</a> Rust errors.
+</body>
+</html>
