+++ src/test/kotlin/org/rust/ide/annotator/RsErrorAnnotatorTest.kt
@@ -425,85 +425,6 @@ class RsErrorAnnotatorTest : RsAnnotatorTestBase() {
         }
     """)
 
-    fun `test E0106 missing lifetime in struct field`() = checkErrors("""
-        struct Foo<'a> {
-            a: &'a str,
-            b: (bool, (u8, &'a f64)),
-            c: Option<Box<&'a u32>>,
-            f: &'a Fn (&u32) -> &u32,
-        }
-        struct Bar<'a> {
-            a: <error descr="Missing lifetime specifier [E0106]">&</error>str,
-            b: (bool, (u8, <error>&</error>f64)),
-            c: Result<Box<<error>&</error>u32>, u8>,
-            f: <error>&</error>Fn (&u32) -> &u32,
-        }
-    """)
-
-    fun `test E0106 missing lifetime in tuple struct field`() = checkErrors("""
-        struct Foo<'a> (
-            &'a str,
-            (bool, (u8, &'a f64)),
-            &'a Fn (&u32) -> &u32);
-        struct Bar<'a> (
-            <error descr="Missing lifetime specifier [E0106]">&</error>str,
-            (bool, (u8, <error>&</error>f64)),
-            <error>&</error>Fn (&u32) -> &u32);
-    """)
-
-    fun `test E0106 missing lifetime in enum`() = checkErrors("""
-        enum Foo<'a> {
-            A(&'a str),
-            B(bool, (u8, &'a f64)),
-            F(&'a Fn (&u32) -> &u32),
-        }
-        enum Bar<'a> {
-            A(<error descr="Missing lifetime specifier [E0106]">&</error>str),
-            B(bool, (u8, <error>&</error>f64)),
-            F(<error>&</error>Fn (&u32) -> &u32),
-        }
-    """)
-
-    fun `test E0106 missing lifetime in type alias`() = checkErrors("""
-        type Str = &'static str;
-        type Foo<'a> = &'a Fn (&u32) -> &u32;
-
-        type U32 = <error descr="Missing lifetime specifier [E0106]">&</error>u32;
-        type Tuple = (bool, (u8, <error>&</error>f64));
-        type Func = <error>&</error>Fn (&u32) -> &u32;
-    """)
-
-    fun `test E0106 missing lifetime ignores raw pointers`() = checkErrors("""
-        struct Foo {
-            raw: *const i32   // Must not be highlighted
-        }
-    """)
-
-    fun `test E0106 missing lifetime in base types`() = checkErrors("""
-        struct Foo1<'a>(&'a str);
-        struct Foo2<'a, 'b> { a: &'a u32, b: &'b str }
-
-        type Err1 = <error descr="Missing lifetime specifier [E0106]">Foo1</error>;
-        struct Err2<'a> { a: <error>Foo2<></error> }
-        enum Err3<'d> { A(&'d Box<<error>Foo1</error>> ) }
-    """)
-
-    fun `test E0107 wrong number of lifetime parameters`() = checkErrors("""
-        struct Foo0;
-        struct Foo1<'a>(&'a str);
-        struct Foo2<'a, 'b> { a: &'a u32, b: &'b str }
-
-        struct Ok<'a> { a: Foo0, b: Foo1<'a>, c: Foo2<'a, 'a> }
-
-        struct Err<'a, 'b, 'c> {
-            a: <error descr="Wrong number of lifetime parameters: expected 0, found 2 [E0107]">Foo0<'a, 'b></error>,
-            b: <error descr="Wrong number of lifetime parameters: expected 1, found 3 [E0107]">Foo1<'a, 'b, 'c></error>,
-            c: <error descr="Wrong number of lifetime parameters: expected 2, found 1 [E0107]">Foo2<'a></error>,
-        }
-        type TErr<'a> = <error>Foo2<'a></error>;
-        enum EErr<'a> { E(Box<<error>Foo1<'a, 'a></error>>) }
-    """)
-
     fun `test E0121 type placeholder in signatures`() = checkErrors("""
         fn ok(_: &'static str) {
             let four = |x: _| 4;
@@ -605,62 +526,6 @@ class RsErrorAnnotatorTest : RsAnnotatorTestBase() {
         impl Foo {
             <error descr="Associated types are not allowed in inherent impls [E0202]">type Long = i64;</error>
         }
-    """)
-
-    fun `test E0243 number of type arguments is less than expected`() = checkErrors("""
-        struct Foo1<T> { t: T }
-        struct Foo2<T, U> { t: T, u: U }
-        struct Foo2to3<T, U, V = bool> { t: T, u: U, v: V }
-
-        struct Ok {
-            ok1: Foo1<u32>,
-            ok2: Foo2<u32, bool>,
-            ok3: Foo2to3<u8, u8>,
-        }
-
-        struct Err {
-            err1: <error descr="Wrong number of type arguments: expected 1, found 0 [E0243]">Foo1</error>,
-            err2: <error descr="Wrong number of type arguments: expected 2, found 1 [E0243]">Foo2<u32></error>,
-            err3: <error descr="Wrong number of type arguments: expected at least 2, found 1 [E0243]">Foo2to3<u32></error>,
-        }
-
-        impl <error>Foo1</error> {}
-        fn err(f: <error>Foo2<u32></error>) -> <error>Foo1</error> {}
-        type Type = <error>Foo2to3<u8></error>;
-    """)
-
-    fun `test E0243 ignores Fn-traits`() = checkErrors("""
-        fn foo(f: &mut FnOnce(u32) -> bool) {}  // No annotation despite the fact that FnOnce has a type parameter
-    """)
-
-    fun `test E0243 ignores Self type`() = checkErrors("""
-        struct Foo<T> { t: T }
-        impl<T> Foo<T> {
-            fn foo(s: Self) {}
-        }
-    """)
-
-    fun `test E0244 number of type arguments is greater than expected`() = checkErrors("""
-        struct Foo0;
-        struct Foo1<T> { t: T }
-        struct Foo1to2<T, U = bool> { t: T, u: U }
-
-        struct Ok {
-            ok1: Foo0,
-            ok2: Foo1<u32>,
-            ok3: Foo1to2<u8>,
-            ok4: Foo1to2<u8, bool>,
-        }
-
-        struct Err {
-            err1: <error descr="Wrong number of type arguments: expected 0, found 2 [E0244]">Foo0<u32, bool></error>,
-            err2: <error descr="Wrong number of type arguments: expected 1, found 2 [E0244]">Foo1<u8, f64></error>,
-            err3: <error descr="Wrong number of type arguments: expected at most 2, found 3 [E0244]">Foo1to2<u32, f32, bool></error>,
-        }
-
-        impl <error>Foo0<u8></error> {}
-        fn err(f: <error>Foo1<u32, bool></error>) -> <error>Foo1to2<u8, u8, u8></error> {}
-        type Type = <error>Foo1<u8, bool, f64></error>;
     """)
 
     fun `test E0261 undeclared lifetimes`() = checkErrors("""
